# pkg/api/docs/swagger.yaml
basePath: /api/v1
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: UDS Runtime API
  version: 0.0.0
paths:
  /monitor/pepr/{stream}:
    get:
      consumes:
        - text/html
      description: Get Pepr data
      parameters:
        - description: stream type to filter on, all streams by default
          enum:
            - AnyStream
            - PolicyStream
            - OperatorStream
            - AllowStream
            - DenyStream
            - MutateStream
            - FailureStream
          in: path
          name: stream
          type: string
      produces:
        - text/event-stream
      responses:
        "200":
          description: OK
      tags:
        - monitor
...
schemes:
  - http
  - https
swagger: "2.0"